upgrading to float

+++ calculator/sub.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
 
-int sub(int a, int b){
+float sub(float a, float b){
     return a-b;
 }
